refactor: 合并 V47-V54 迁移，消除字段反复
将 V47-V54 的 8 次迁移压缩为 1 次合并迁移 (v47_v53.rs)：
- V47: 建 blackboards 表时一次性包含所有演进字段
  (pending_record_ids / debounce_secs / debounce_count /
   wiki_index_prompt / wiki_page_prompt)
- V48: 修复 todos 索引为 workspace 级
- V53: 建 blackboard_pages 表

删除 V49/V51/V52 (pending_todo_ids 加了就删)
删除 V50 (字段已并入 V47)
删除 V54 (字段已并入 V47)

最终迁移链: 19 → 12 个, 零 ADD/DROP COLUMN 反复
