+++ backend/src/db/migration/mod.rs
@@ -11,14 +11,7 @@ use super::Database;
 mod v1;
 mod v2_v5;
 mod v41_v46;
-mod v47;
-mod v48;
-mod v49;
-mod v50;
-mod v51;
-mod v52;
-mod v53;
-mod v54;
+mod v47_v53;
 
 pub use v2_v5::read_applied_versions;
 pub use v2_v5::drop_column_if_exists;
@@ -52,14 +45,7 @@ pub(super) fn all_migrations() -> Vec<Box<dyn Migration>> {
         Box::new(v41_v46::V44AddFeishuMessagesProcessedId),
         Box::new(v41_v46::V45AddTodosActionType),
         Box::new(v41_v46::V46AddTodosActionKey),
-        Box::new(v47::V47CreateBlackboardsTable),
-        Box::new(v48::V48ScopeTodosActionKeyByWorkspace),
-        Box::new(v49::V49AddBlackboardPendingTodoIds),
-        Box::new(v50::V50AddBlackboardWorkspaceConfig),
-        Box::new(v51::V51AddBlackboardPendingRecordIds),
-        Box::new(v52::V52DropBlackboardPendingTodoIds),
-        Box::new(v53::V53CreateBlackboardPagesTable),
-        Box::new(v54::V54SplitBlackboardPrompt),
+        Box::new(v47_v53::V47ConsolidatedBlackboardFeatures),
     ]
 }
